# backend/openapi.yaml
openapi: 3.1.0
info:
    title: VidyaSetu
    version: 0.1.0
    description: |
        VidyaSetu is a role-based education and workshop management backend API designed for Admin, Institutional Admin, Educator, Student, and Public users.

        The system supports the complete academic lifecycle including authentication, content delivery, assessments, certification, analytics, communication, and AI-powered features.

        #### Core capabilities include:

        - **Secure JWT-based authentication** and role-based access control for admin, institutional admin, educator, and student users.
        - **Institution and workshop administration**, including workshop creation, updates, enrollment tracking, assignments, and lifecycle management.
        - **Centralized content management** for modules, study materials, and learning resources uploaded by educators and accessed by students.
        - **Assessment management** for quizzes, tests, and assignments, including question creation, student attempts, answer submission, and result storage.
        - **Automated evaluation support** for objective assessments to reduce manual grading effort and improve consistency.
        - **Certificate generation** for successful workshop completion and public certificate verification using unique verification codes or identifiers.
        - **Performance tracking and reporting** for student progress, workshop statistics, assessment outcomes, completion rates, and trend analysis.
...

...

...

...

...

...

...

...

...
components:
    schemas:
...
        HTTPValidationError:
            properties:
                detail:
                    items:
                        $ref: "#/components/schemas/ValidationError"
                    title: Detail
                    type: array
            title: HTTPValidationError
            type: object
...
        InstitutionCreate:
            properties:
                address:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Address
                admin_id:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Admin Id
                name:
                    title: Name
                    type: string
            required:
                - name
            title: InstitutionCreate
            type: object
...
        InstitutionResponse:
            properties:
                address:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Address
                admin_id:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Admin Id
                id:
                    title: Id
                    type: string
                name:
                    title: Name
                    type: string
            required:
                - id
                - name
                - address
                - admin_id
            title: InstitutionResponse
            type: object
...
        InstitutionUpdate:
            properties:
                address:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Address
                admin_id:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Admin Id
                name:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Name
            title: InstitutionUpdate
            type: object
...
        MaterialDownloadResponse:
            properties:
                download_url:
                    title: Download Url
                    type: string
                material_id:
                    title: Material Id
                    type: string
                module_id:
                    title: Module Id
                    type: string
            required:
                - module_id
                - material_id
                - download_url
            title: MaterialDownloadResponse
            type: object
        MaterialItem:
            properties:
                content:
                    title: Content
                    type: string
                created_at:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Created At
                id:
                    title: Id
                    type: string
                title:
                    title: Title
                    type: string
                type:
                    enum:
                        - video
                        - text
                        - link
                    title: Type
                    type: string
            required:
                - id
                - title
                - type
                - content
            title: MaterialItem
            type: object
        MaterialItemUpdate:
            description: Partial update schema for a single material embedded in Module.materials.
            properties:
                content:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Content
                title:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Title
                type:
                    anyOf:
                        - enum:
                              - video
                              - text
                              - link
                          type: string
                        - type: "null"
                    title: Type
            title: MaterialItemUpdate
            type: object
...
        ModuleResponse:
            properties:
                id:
                    title: Id
                    type: string
                materials:
                    default: []
                    items:
                        $ref: "#/components/schemas/MaterialItem"
                    title: Materials
                    type: array
                order_index:
                    anyOf:
                        - type: integer
                        - type: "null"
                    title: Order Index
                title:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Title
                workshop_id:
                    anyOf:
                        - type: string
                        - type: "null"
                    title: Workshop Id
            required:
                - id
                - workshop_id
                - title
                - order_index
            title: ModuleResponse
            type: object
...
        ValidationError:
            properties:
                ctx:
                    title: Context
                    type: object
                input:
                    title: Input
                loc:
                    items:
                        anyOf:
                            - type: string
                            - type: integer
                    title: Location
                    type: array
                msg:
                    title: Message
                    type: string
                type:
                    title: Error Type
                    type: string
            required:
                - loc
                - msg
                - type
            title: ValidationError
            type: object
...
    securitySchemes:
        OAuth2PasswordBearer:
            flows:
                password:
                    scopes: {}
                    tokenUrl: /api/v1/auth/login
            type: oauth2
paths:
...
    /api/v1/institutions/:
        get:
            operationId: list_institutions_api_v1_institutions__get
            parameters:
                - in: query
                  name: offset
                  required: false
                  schema:
                      default: 0
                      title: Offset
                      type: integer
                - in: query
                  name: limit
                  required: false
                  schema:
                      default: 50
                      title: Limit
                      type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/InstitutionResponse"
                                title: Response List Institutions Api V1 Institutions  Get
                                type: array
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: List institutions
            tags:
                - institutions
        post:
            operationId: create_api_v1_institutions__post
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/InstitutionCreate"
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/InstitutionResponse"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Create institution (admin only)
            tags:
                - institutions
    /api/v1/institutions/{institution_id}:
        delete:
            operationId: delete_one_api_v1_institutions__institution_id__delete
            parameters:
                - in: path
                  name: institution_id
                  required: true
                  schema:
                      title: Institution Id
                      type: string
            responses:
                "204":
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Delete institution (admin only)
            tags:
                - institutions
        get:
            operationId: get_one_api_v1_institutions__institution_id__get
            parameters:
                - in: path
                  name: institution_id
                  required: true
                  schema:
                      title: Institution Id
                      type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/InstitutionResponse"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Get institution by ID
            tags:
                - institutions
        patch:
            operationId: update_one_api_v1_institutions__institution_id__patch
            parameters:
                - in: path
                  name: institution_id
                  required: true
                  schema:
                      title: Institution Id
                      type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/InstitutionUpdate"
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/InstitutionResponse"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Update institution (admin only)
            tags:
                - institutions
    /api/v1/materials/{module_id}/{material_id}:
        delete:
            operationId: delete_material_api_v1_materials__module_id___material_id__delete
            parameters:
                - in: path
                  name: module_id
                  required: true
                  schema:
                      title: Module Id
                      type: string
                - in: path
                  name: material_id
                  required: true
                  schema:
                      title: Material Id
                      type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ModuleResponse"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Remove a single material item from a module (staff only)
            tags:
                - materials
        patch:
            operationId: patch_material_api_v1_materials__module_id___material_id__patch
            parameters:
                - in: path
                  name: module_id
                  required: true
                  schema:
                      title: Module Id
                      type: string
                - in: path
                  name: material_id
                  required: true
                  schema:
                      title: Material Id
                      type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/MaterialItemUpdate"
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ModuleResponse"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Update a single material item inside a module (staff only)
            tags:
                - materials
    /api/v1/materials/{module_id}/{material_id}/download:
        get:
            operationId: download_material_api_v1_materials__module_id___material_id__download_get
            parameters:
                - in: path
                  name: module_id
                  required: true
                  schema:
                      title: Module Id
                      type: string
                - in: path
                  name: material_id
                  required: true
                  schema:
                      title: Material Id
                      type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/MaterialDownloadResponse"
                    description: Successful Response
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/HTTPValidationError"
                    description: Validation Error
            security:
                - OAuth2PasswordBearer: []
            summary: Get a download URL for one module material
            tags:
                - materials
...